Add files via upload
This is my Second Publication with Annapoorna PK (Corresponding Author) and The Mind Gala (a joint initiative between Project Encephelon and Thakur Neurodegeneration Lab) in 'Tales of Neuroscience' which is a collection of articles written by students from across India. This book covers the basics of neuroscience, the impact of neuroscience on our daily lives, mental health disorders, and the latest developments in neuroscience and technology. 
         In this publication, I have highlighted the neuropathology of Naegleria fowleri, a tropical Amoeba which is notoriously famous as "The Brain Eater" and its infection in the brain. It explores the immunological response of the human body toward an Amoebic infection and the effects of the infection seen in patients.